# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: The Ministry of Herbology API
  version: 1.0.0
  description: |
    Contract for the Ministry of Herbology. Frozen in S0 — only Workstream A
    changes this file, via an ADR. Every path declares the workstream that
    implements it with `x-workstream`; `tests/contract/` enforces that the
    running API matches this document.

    Units: millimetres, °C, litres, lux. Timestamps are RFC 3339 UTC.
servers:
  - url: /api/v1
...
paths:
...
  /specimens:
    get:
      tags: [inventory]
      x-workstream: C
      summary: List and search specimens
      operationId: listSpecimens
      parameters:
        - { name: q,           in: query, schema: { type: string } }
        - { name: location_id, in: query, schema: { type: string, format: uuid } }
        - { name: outdoor,     in: query, schema: { type: boolean } }
        - { name: status,      in: query, schema: { $ref: '#/components/schemas/SpecimenStatus' } }
        - { name: toxic_to_pets, in: query, schema: { type: boolean } }
        - { name: limit,  in: query, schema: { type: integer, default: 50, maximum: 200 } }
        - { name: cursor, in: query, schema: { type: string } }
      responses:
        '200':
          description: A page of specimens
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SpecimenPage' }
    post:
      tags: [inventory]
      x-workstream: C
      summary: Add a plant. Enrichment is queued and runs in the background.
      operationId: createSpecimen
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SpecimenCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Specimen' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /specimens/{specimen_id}:
    parameters:
      - { name: specimen_id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [inventory]
      x-workstream: C
      operationId: getSpecimen
      responses:
        '200':
          description: One specimen
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Specimen' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [inventory]
      x-workstream: C
      operationId: updateSpecimen
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SpecimenUpdate' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Specimen' }
    delete:
      tags: [inventory]
      x-workstream: C
      summary: Archive a specimen (soft delete)
      operationId: archiveSpecimen
      responses:
        '204': { description: Archived }
  /specimens/{specimen_id}/photos:
    parameters:
      - { name: specimen_id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [inventory]
      x-workstream: C
      operationId: listPhotos
      responses:
        '200':
          description: Growth log photos, newest first
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Photo' }
    post:
      tags: [inventory]
      x-workstream: C
      operationId: uploadPhoto
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:    { type: string, format: binary }
                caption: { type: string }
      responses:
        '201':
          description: Uploaded
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Photo' }
  /specimens/{specimen_id}/log:
    parameters:
      - { name: specimen_id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [inventory]
      x-workstream: C
      operationId: listLogEntries
      responses:
        '200':
          description: Log entries, newest first
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/LogEntry' }
    post:
      tags: [inventory]
      x-workstream: C
      operationId: createLogEntry
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LogEntryCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LogEntry' }
  /locations:
    get:
      tags: [inventory]
      x-workstream: C
      operationId: listLocations
      parameters:
        - { name: site_id, in: query, schema: { type: string, format: uuid } }
        - { name: outdoor, in: query, schema: { type: boolean } }
      responses:
        '200':
          description: The site → area → zone tree, flattened
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Location' }
    post:
      tags: [inventory]
      x-workstream: C
      operationId: createLocation
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LocationCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Location' }
...
  /species:
    get:
      tags: [botany]
      x-workstream: D
      operationId: listSpecies
      parameters:
        - { name: q, in: query, schema: { type: string } }
      responses:
        '200':
          description: Known species
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Species' }
  /species/{species_id}:
    parameters:
      - { name: species_id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [botany]
      x-workstream: D
      operationId: getSpecies
      responses:
        '200':
          description: Species with its cited care profile
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Species' }
  /species/{species_id}/care-values:
    parameters:
      - { name: species_id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [botany]
      x-workstream: D
      summary: Every care field with its source and confidence
      operationId: listCareValues
      responses:
        '200':
          description: Cited care values
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/CareValue' }
    put:
      tags: [botany]
      x-workstream: D
      summary: User override of a care value
      operationId: overrideCareValue
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CareValueOverride' }
      responses:
        '200':
          description: Stored override
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CareValue' }
...
  /almanac/forecast:
    get:
      tags: [almanac]
      x-workstream: E
      summary: 1-day (hourly) and 10-day (daily) forecast
      operationId: getForecast
      parameters:
        - { name: site_id, in: query, required: true, schema: { type: string, format: uuid } }
        - name: horizon
          in: query
          schema: { type: string, enum: [hourly, daily], default: daily }
      responses:
        '200':
          description: Forecast series
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ForecastPoint' }
  /almanac/history:
    get:
      tags: [almanac]
      x-workstream: E
      summary: 1 / 7 / 30-day indoor and outdoor conditions history
      operationId: getHistory
      parameters:
        - name: window
          in: query
          required: true
          schema: { type: string, enum: ['1d', '7d', '30d'] }
        - { name: site_id,     in: query, schema: { type: string, format: uuid } }
        - { name: location_id, in: query, schema: { type: string, format: uuid } }
        - { name: specimen_id, in: query, schema: { type: string, format: uuid } }
        - name: metric
          in: query
          schema:
            type: string
            enum: [temperature_c, humidity_pct, soil_moisture_pct, precip_mm, et0_mm]
      responses:
        '200':
          description: Bucketed series ready for uPlot
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Series' }
  /almanac/water-balance/{specimen_id}:
    parameters:
      - { name: specimen_id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [almanac]
      x-workstream: E
      summary: Soil water deficit history and current state
      operationId: getWaterBalance
      responses:
        '200':
          description: Water balance
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WaterBalance' }
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: moh_session
  responses:
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [detail]
      properties:
        detail: { type: string }
        code:   { type: string }
    Confidence:
      type: string
      enum: [high, medium, low, unknown]
    SpecimenStatus:
      type: string
      enum: [thriving, struggling, dormant, overwintering, lost, given_away, archived]
    TaskStatus:
      type: string
      enum: [due, done, satisfied, skipped, cancelled]
    TaskType:
      type: string
      enum: [water, fertilize, mist, prune, repot, inspect,
             bring_indoors, return_outdoors, cover, rotate]
    SunExposure:
      type: string
      enum: [full_sun, part_sun, part_shade, full_shade, unknown]
...
    Location:
      type: object
      required: [id, name, kind, is_outdoor]
      properties:
        id:           { type: string, format: uuid }
        site_id:      { type: string, format: uuid }
        parent_id:    { type: [string, 'null'], format: uuid }
        name:         { type: string }
        kind:         { type: string, enum: [area, zone, bed, room, shelf] }
        is_outdoor:   { type: boolean }
        is_covered:   { type: boolean }
        sun_exposure: { $ref: '#/components/schemas/SunExposure' }
        map_layer_id: { type: [string, 'null'], format: uuid }
        boundary_px:
          type: [array, 'null']
          items:
            type: array
            items: { type: number }
        specimen_count: { type: integer, readOnly: true }
    LocationCreate:
      type: object
      required: [name, kind, is_outdoor]
      properties:
        site_id:      { type: string, format: uuid }
        parent_id:    { type: [string, 'null'], format: uuid }
        name:         { type: string }
        kind:         { type: string, enum: [area, zone, bed, room, shelf] }
        is_outdoor:   { type: boolean }
        is_covered:   { type: boolean, default: false }
        sun_exposure: { $ref: '#/components/schemas/SunExposure' }
    Specimen:
      type: object
      required: [id, is_outdoor, status, display_name]
      properties:
        id:              { type: string, format: uuid }
        display_name:    { type: string, description: 'nickname, else common name, else accepted name' }
        nickname:        { type: [string, 'null'] }
        cultivar:        { type: [string, 'null'] }
        species:         { $ref: '#/components/schemas/SpeciesBrief' }
        is_group:        { type: boolean }
        count:           { type: integer }
        location:        { $ref: '#/components/schemas/Location' }
        map_layer_id:    { type: [string, 'null'], format: uuid }
        pin_px:          { $ref: '#/components/schemas/PixelPoint' }
        is_outdoor:      { type: boolean }
        in_container:    { type: boolean }
        container_litres: { type: [number, 'null'] }
        soil_note:       { type: [string, 'null'] }
        acquired_on:     { type: [string, 'null'], format: date }
        provenance:      { type: [string, 'null'] }
        status:          { $ref: '#/components/schemas/SpecimenStatus' }
        primary_photo_url: { type: [string, 'null'] }
        next_task:       { $ref: '#/components/schemas/TaskBrief' }
        created_at:      { type: string, format: date-time }
...
    SpecimenCreate:
      type: object
      required: [name]
      properties:
        name:         { type: string, description: Typed name; resolved to a species }
        species_id:   { type: string, format: uuid, description: Skip resolution }
        image_key:    { type: string, description: Photo for Pl@ntNet identification }
        nickname:     { type: string }
        cultivar:     { type: string }
        location_id:  { type: string, format: uuid }
        in_container: { type: boolean, default: true }
        container_litres: { type: number }
        count:        { type: integer, default: 1 }
        acquired_on:  { type: string, format: date }
    SpecimenUpdate:
      type: object
      properties:
        nickname:        { type: [string, 'null'] }
        cultivar:        { type: [string, 'null'] }
        location_id:     { type: [string, 'null'], format: uuid }
        in_container:    { type: boolean }
        container_litres: { type: [number, 'null'] }
        soil_note:       { type: [string, 'null'] }
        status:          { $ref: '#/components/schemas/SpecimenStatus' }
        notes:           { type: [string, 'null'] }
        water_k_c_override:           { type: [number, 'null'] }
        water_interval_days_override: { type: [integer, 'null'] }
        min_temp_c_override:          { type: [number, 'null'] }
    SpecimenPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Specimen' }
        next_cursor: { type: [string, 'null'] }
        total:       { type: integer }
    SpeciesBrief:
      type: object
      properties:
        id:            { type: string, format: uuid }
        accepted_name: { type: string }
        common_name:   { type: [string, 'null'] }
        family:        { type: [string, 'null'] }
    Species:
      allOf:
        - $ref: '#/components/schemas/SpeciesBrief'
        - type: object
          properties:
            common_names:     { type: array, items: { type: string } }
            genus:            { type: [string, 'null'] }
            native_range:     { type: array, items: { type: string } }
            summary:          { type: [string, 'null'] }
            light_label:      { type: [string, 'null'] }
            water_k_c:        { type: [number, 'null'] }
            water_interval_days: { type: [integer, 'null'] }
            soil_ph_min:      { type: [number, 'null'] }
            soil_ph_max:      { type: [number, 'null'] }
            fertilizer_note:  { type: [string, 'null'] }
            humidity_min_pct: { type: [integer, 'null'] }
            min_temp_c:       { type: [number, 'null'] }
            usda_zone_min:    { type: [string, 'null'] }
            usda_zone_max:    { type: [string, 'null'] }
            dormancy_months:  { type: array, items: { type: integer } }
            toxic_to_pets:     { type: [boolean, 'null'] }
            toxic_to_children: { type: [boolean, 'null'] }
            toxicity_note:    { type: [string, 'null'] }
            enrichment_state: { type: string, enum: [pending, running, complete, failed] }
            sources:
              type: array
              items: { $ref: '#/components/schemas/SourceRef' }
    SourceRef:
      type: object
      required: [id, kind, title]
      properties:
        id:      { type: string, format: uuid }
        kind:    { type: string }
        title:   { type: string }
        url:     { type: [string, 'null'] }
        license: { type: [string, 'null'] }
        retrieved_at: { type: string, format: date-time }
    CareValue:
      type: object
      required: [field, value, confidence]
      description: |
        Rule 6 in the working agreement — no invented plant facts. A value with
        no `source` must carry `confidence: unknown` unless it is a user override.
      properties:
        id:         { type: string, format: uuid }
        field:      { type: string, example: water_k_c }
        value:      {}
        unit:       { type: [string, 'null'] }
        source:     { $ref: '#/components/schemas/SourceRef' }
        confidence: { $ref: '#/components/schemas/Confidence' }
        is_user_override: { type: boolean }
        note:       { type: [string, 'null'] }
    CareValueOverride:
      type: object
      required: [field, value]
      properties:
        field: { type: string }
        value: {}
        unit:  { type: string }
        note:  { type: string }
...
    ForecastPoint:
      type: object
      required: [time]
      properties:
        time:            { type: string, format: date-time }
        temperature_c:   { type: [number, 'null'] }
        temp_min_c:      { type: [number, 'null'] }
        temp_max_c:      { type: [number, 'null'] }
        precip_mm:       { type: [number, 'null'] }
        precip_prob_pct: { type: [number, 'null'] }
        et0_mm:          { type: [number, 'null'] }
        wind_kph:        { type: [number, 'null'] }
        condition:       { type: [string, 'null'] }
        sunrise:         { type: [string, 'null'], format: date-time }
        sunset:          { type: [string, 'null'], format: date-time }
    Series:
      type: object
      description: Column-oriented, ready to hand straight to uPlot.
      required: [metric, times, values]
      properties:
        metric: { type: string }
        unit:   { type: string }
        times:  { type: array, items: { type: integer }, description: Unix seconds }
        values: { type: array, items: { type: [number, 'null'] } }
        min:    { type: array, items: { type: [number, 'null'] } }
        max:    { type: array, items: { type: [number, 'null'] } }
        label:  { type: string }
    WaterBalance:
      type: object
      required: [specimen_id, deficit_mm, threshold_mm, is_due]
      properties:
        specimen_id:  { type: string, format: uuid }
        deficit_mm:   { type: number }
        capacity_mm:  { type: number }
        threshold_mm: { type: number }
        k_c:          { type: number }
        is_due:       { type: boolean }
        sensor_override_pct: { type: [number, 'null'] }
        days:
          type: array
          items:
            type: object
            properties:
              day:           { type: string, format: date }
              deficit_mm:    { type: number }
              et0_mm:        { type: number }
              precip_mm:     { type: number }
              irrigation_mm: { type: number }
...
    TaskBrief:
      type: object
      properties:
        id:          { type: string, format: uuid }
        task_type:   { $ref: '#/components/schemas/TaskType' }
        due_at:      { type: string, format: date-time }
        status:      { $ref: '#/components/schemas/TaskStatus' }
        plain_title: { type: string }
...
    PixelPoint:
      type: [object, 'null']
      properties:
        x: { type: number }
        y: { type: number }
...
    Photo:
      type: object
      required: [id, url, taken_at]
      properties:
        id:         { type: string, format: uuid }
        url:        { type: string }
        thumb_url:  { type: [string, 'null'] }
        taken_at:   { type: string, format: date-time }
        caption:    { type: [string, 'null'] }
        is_primary: { type: boolean }
    LogEntry:
      type: object
      required: [id, kind, occurred_at]
      properties:
        id:          { type: string, format: uuid }
        kind:        { type: string, enum: [growth, pest, disease, repot, prune, relocate, note] }
        occurred_at: { type: string, format: date-time }
        body:        { type: [string, 'null'] }
        photo:       { $ref: '#/components/schemas/Photo' }
        data:        { type: object, additionalProperties: true }
    LogEntryCreate:
      type: object
      required: [kind]
      properties:
        kind:        { type: string, enum: [growth, pest, disease, repot, prune, relocate, note] }
        body:        { type: string }
        occurred_at: { type: string, format: date-time }
        photo_id:    { type: string, format: uuid }
...
security:
  - session: []